# root/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/meters/{id}": {
      "get": {
        "operationId": "MetersController_findOne",
        "summary": "Get Meter by ID",
        "description": "Retrieves a single meter record by its UUID including building_id, meter_code, meter_type, zone, and status. The Organization Admin, Technician, and Sustainability Officer can look up a specific meter. Pass the meter_id as a URL path parameter.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter record returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Meter with the given ID not found."
          }
        },
        "tags": [
          "meters"
        ]
      },
      "put": {
        "operationId": "MetersController_put",
        "summary": "Replace Meter",
        "description": "Completely replaces an existing meter record with new data. Only the Organization Admin can perform full meter replacements. Send a PUT request with the meter_id in the URL and a complete PutMeterDto JSON body.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutMeterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meter record replaced successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin."
          },
          "404": {
            "description": "Meter with the given ID not found."
          }
        },
        "tags": [
          "meters"
        ]
      },
      "patch": {
        "operationId": "MetersController_update",
        "summary": "Update Meter",
        "description": "Partially updates specific fields of an existing meter (e.g., status from active to faulty, or zone reassignment). Only the Organization Admin can modify meter configuration. Send a PATCH request with the meter_id in the URL and an UpdateMeterDto JSON body containing only the fields to change.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meter record updated successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin."
          },
          "404": {
            "description": "Meter with the given ID not found."
          }
        },
        "tags": [
          "meters"
        ]
      },
      "delete": {
        "operationId": "MetersController_remove",
        "summary": "Delete Meter",
        "description": "Permanently removes a meter record from the system. Only the Organization Admin can decommission and delete meters. Pass the meter_id as a URL path parameter.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter record deleted successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin."
          },
          "404": {
            "description": "Meter with the given ID not found."
          }
        },
        "tags": [
          "meters"
        ]
      }
    },
...
    "/api/meter-readings": {
      "post": {
        "operationId": "MeterReadingsController_create",
        "summary": "Create Meter Reading",
        "description": "Records a new meter reading from an IoT sensor. The Organization Admin and Technician can submit readings by providing meter_id, value, and unit. Send a POST request with a CreateMeterReadingDto JSON body.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeterReadingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meter reading created successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin or Technician."
          }
        },
        "tags": [
          "meter-readings"
        ]
      },
      "get": {
        "operationId": "MeterReadingsController_findAll",
        "summary": "List All Meter Readings",
        "description": "Retrieves all meter reading records across the system. The Organization Admin, Technician, Financial Analyst, and Sustainability Officer can view the full history of sensor readings. Send a GET request with no additional parameters.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of all meter reading records returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          }
        },
        "tags": [
          "meter-readings"
        ]
      }
    },
    "/api/meter-readings/upload": {
      "post": {
        "operationId": "MeterReadingsController_uploadSpreadsheet",
        "summary": "Bulk Upload Meter Readings",
        "description": "Uploads a CSV file of meter readings for sites without automatic sensors. The Organization Admin and Technician can submit a batch of readings via file. Send a multipart/form-data POST request with the CSV file under the 'file' field.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meter readings imported successfully from CSV."
          },
          "400": {
            "description": "File missing, wrong type, or over the size limit."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin or Technician."
          }
        },
        "tags": [
          "meter-readings"
        ]
      }
    },
    "/api/meter-readings/{id}": {
      "get": {
        "operationId": "MeterReadingsController_findOne",
        "summary": "Get Meter Reading by ID",
        "description": "Retrieves a single meter reading record by its UUID including meter_id, value, unit, and timestamp. The Organization Admin, Technician, Financial Analyst, and Sustainability Officer can look up a specific reading. Pass the reading_id as a URL path parameter.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter reading record returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Meter reading with the given ID not found."
          }
        },
        "tags": [
          "meter-readings"
        ]
      },
      "put": {
        "operationId": "MeterReadingsController_put",
        "summary": "Replace Meter Reading",
        "description": "Completely replaces an existing meter reading record with new data. Only the Organization Admin can perform full reading replacements. Send a PUT request with the reading_id in the URL and a complete PutMeterReadingDto JSON body.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutMeterReadingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meter reading record replaced successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin."
          },
          "404": {
            "description": "Meter reading with the given ID not found."
          }
        },
        "tags": [
          "meter-readings"
        ]
      },
      "patch": {
        "operationId": "MeterReadingsController_update",
        "summary": "Update Meter Reading",
        "description": "Partially updates specific fields of an existing meter reading (e.g., correcting a value or unit). Only the Organization Admin can modify readings. Send a PATCH request with the reading_id in the URL and an UpdateMeterReadingDto JSON body containing only the fields to change.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeterReadingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meter reading record updated successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin."
          },
          "404": {
            "description": "Meter reading with the given ID not found."
          }
        },
        "tags": [
          "meter-readings"
        ]
      },
      "delete": {
        "operationId": "MeterReadingsController_remove",
        "summary": "Delete Meter Reading",
        "description": "Permanently removes a meter reading record from the system. Only the Organization Admin can delete readings. Pass the reading_id as a URL path parameter.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC. Enum: Organization Admin | Financial Analyst | Technician | Sustainability Officer | Campus Visitor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter reading record deleted successfully."
          },
          "403": {
            "description": "Forbidden – caller role is not Organization Admin."
          },
          "404": {
            "description": "Meter reading with the given ID not found."
          }
        },
        "tags": [
          "meter-readings"
        ]
      }
    },
...
  },
  "info": {
    "title": "EnerTrack API",
    "description": "Backend API for EnerTrack — Facility Management, Sustainability & Finance Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      },
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id"
      }
    },
    "schemas": {
...
      "PutMeterDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMeterDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "building_id": "660e8800-0002-4000-8000-000000000000",
          "meter_code": "M-001",
          "meter_type": "electricity",
          "zone": "Zone 2",
          "status": "active"
        }
      },
      "CreateMeterReadingDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "meter_id": "mmmm0000-0002-4000-8000-000000000000",
          "value": 105.5,
          "unit": "unit",
          "timestamp": "2025-01-01T10:00:00.000Z"
        }
      },
      "PutMeterReadingDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMeterReadingDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "meter_id": "mmmm0000-0002-4000-8000-000000000000",
          "value": 105.5,
          "unit": "unit",
          "timestamp": "2025-01-01T10:00:00.000Z"
        }
      },
...
    }
  }
}